refactor SchedulerManagerImpl -> SchedulerManager
OCD-3119
